The game scene logic is now in the GameScene not in the AbstractGameScene. The level is passed to the game scene so 1 scene will rule them all.
