+++ src/main/java/com/fabbroniko/gameobjects/Enemy.java
@@ -5,23 +5,14 @@ import com.fabbroniko.environment.CollisionDirection;
 import com.fabbroniko.environment.ObjectType;
 import com.fabbroniko.environment.TileMap;
 import com.fabbroniko.environment.Animation.AnimationListener;
-import com.fabbroniko.gamestatemanager.AbstractGenericLevel;
+import com.fabbroniko.scene.GameScene;
 
-/**
- * Represents an enemy that can only be killed if a player hits him from the top.
- * @author com.fabbroniko
- */
 public class Enemy extends AbstractGameObject implements AnimationListener {
 	
 	private boolean init;
-	
-	/**
-	 * Constructs a new Enemy.
-	 * @param tileMap Reference of the {@link TileMap TileMap} on which it should be placed.
-	 * @param level Reference of the {@link AbstractGenericLevel AbstractGenericLevel} on which it should be placed.
-	 */
-	public Enemy(final TileMap tileMap, final AbstractGenericLevel level, final Integer objectID) {
-		super(tileMap, level, Animations.ENEMY_WALK, objectID); 
+
+	public Enemy(final TileMap tileMap, final GameScene gameScene, final Integer objectID) {
+		super(tileMap, gameScene, Animations.ENEMY_WALK, objectID);
 		falling = true;
 		leftOffset = -1;
 		rightOffset = 1;
@@ -56,7 +47,7 @@ public class Enemy extends AbstractGameObject implements AnimationListener {
 		if (direction.equals(CollisionDirection.TOP_COLLISION) && obj.getObjectType().equals(ObjectType.TYPE_PLAYER) && !currentAnimation.getAnimation().equals(Animations.ENEMY_DEAD)) {
 			this.setAnimation(Animations.ENEMY_DEAD);
 			this.currentAnimation.setAnimationListener(this);
-			this.level.getAudioManager().playEffect("hit");
+			this.gameScene.getAudioManager().playEffect("hit");
 		}
 	}
