+++ src/main/java/com/fabbroniko/gameobjects/FallingBlock.java
@@ -4,22 +4,12 @@ import com.fabbroniko.environment.Animations;
 import com.fabbroniko.environment.CollisionDirection;
 import com.fabbroniko.environment.ObjectType;
 import com.fabbroniko.environment.TileMap;
-import com.fabbroniko.gamestatemanager.AbstractGenericLevel;
+import com.fabbroniko.scene.GameScene;
 
-/**
- * A sequence of blocks that start falling when a player hit them from the top.
- * @author com.fabbroniko
- */
 public class FallingBlock extends AbstractGameObject {
-	
-	
-	/**
-	 * Constructs a new FallingBlock.
-	 * @param tileMap Reference of the {@link TileMap TileMap} on which it should be placed.
-	 * @param level Reference of the {@link AbstractGenericLevel AbstractGenericLevel} on which it should be placed.
-	 */
-	public FallingBlock(final TileMap tileMap, final AbstractGenericLevel level, final Integer objectID) {
-		super(tileMap, level, Animations.FALLING_BLOCK, objectID);
+
+	public FallingBlock(final TileMap tileMap, final GameScene gameScene, final Integer objectID) {
+		super(tileMap, gameScene, Animations.FALLING_BLOCK, objectID);
 	}
 	
 	@Override
